requirements.txt
git+https://github.com/huggingface/transformers
datasets
monai
torch
torchvision
numpy
Pillow
matplotlib
tqdm